Improve change membership fee structure
- Rename entry point and rood component
- Split messages into separate components
- Remove nesting from state if statements
- Fix tests
- Improve some variable names

Ticket: https://phabricator.wikimedia.org/T396008
